% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelPlot.R
\name{modelPlot}
\alias{modelPlot}
\title{Plot a function as a line}
\usage{
modelPlot(title = "", x, y, xlim, ylim, ...)
}
\arguments{
\item{title:}{the title of the plot}

\item{x:}{values on the x axis.}

\item{y:}{values on the y axis.}

\item{xlim:}{a vector containing the lower and the upper limit on the x axis}

\item{ylim:}{a vector containing the lower and the upper limit on the y axis}

\item{...:}{other arguments to pass to the lower functions}
}
\value{
nothing: just perform the plot
}
\description{
Plot a bidimensional line plot using the given parameter.
The title and other arguments may also be passed to te underlying plot function.
}
\seealso{
plot3D
}

