% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Plot a bidimensional scatter plot}
\usage{
scatterPlot(title = "", x, y, xlim = c(min(x), max(x)), ylim = c(min(y),
  max(y)), ...)
}
\arguments{
\item{title:}{the title of the plot. Default: nothing}

\item{x:}{values on the x axis.}

\item{y:}{values on the y axis.}

\item{xlim:}{vector of limits on the x axis. Default: min and max of x.}

\item{ylim:}{vector of limits on the x axis. Default: min and max of y.}

\item{...:}{other arguments to pass to the lower functions}
}
\value{
nothing: just perform the plot
}
\description{
Plot a bidimensional scatter plot using the given parameter.
The title and other arguments may also be passed to te underlying plot function.
#' The plot is performed in the maximum available region.
}
\seealso{
plot3D
}

