% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Plot a bidimensional scatter plot}
\usage{
scatterPlot(title = "", x, y, ...)
}
\arguments{
\item{title:}{the title of the plot}

\item{x:}{values on the x axis.}

\item{y:}{values on the y axis.}

\item{...:}{other arguments to pass to the lower functions}
}
\value{
nothing: just perform the plot
}
\description{
Plot a bidimensional scatter plot using the given parameter.
The title and other arguments may also be passed to te underlying plot function.
}
\seealso{
plot3D
}

